+++ src/LaravelShortcutsServiceProvider.php
@@ -4,6 +4,8 @@ namespace Bisual\LaravelShortcuts;
 
 use Bisual\LaravelShortcuts\Commands\BisualResourceMakeCommand;
 use Bisual\LaravelShortcuts\Commands\DtoMakeCommand;
+use Bisual\LaravelShortcuts\Commands\McpCrudResourceMakeCommand;
+use Bisual\LaravelShortcuts\Commands\McpToolMakeCommand;
 use Bisual\LaravelShortcuts\Commands\RepositoryMakeCommand;
 use Bisual\LaravelShortcuts\Commands\RequestDtoMakeCommand;
 use Spatie\LaravelPackageTools\Package;
@@ -25,6 +27,8 @@ class LaravelShortcutsServiceProvider extends PackageServiceProvider
                 DtoMakeCommand::class,
                 RequestDtoMakeCommand::class,
                 BisualResourceMakeCommand::class,
+                McpCrudResourceMakeCommand::class,
+                McpToolMakeCommand::class,
             );
     }
 }